# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Penguin Statistics v3 API, re-designed to aim for lightweight on wire.",
        "title": "Penguin Statistics API",
        "contact": {
            "name": "AlvISs_Reimu",
            "url": "https://github.com/AlvISsReimu",
            "email": "alvissreimu@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "3.0.0-alpha.1"
    },
    "host": "https://penguin-stats.io",
    "basePath": "/api",
    "paths": {
...
        "/PenguinStats/api/v2/zones": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zone"
                ],
                "summary": "Get All Zones",
                "deprecated": true,
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "allOf": [
                                    {
                                        "$ref": "#/definitions/shims.Zone"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "existence": {
                                                "$ref": "#/definitions/models.Existence"
                                            },
                                            "zoneName_i18n": {
                                                "$ref": "#/definitions/models.I18nString"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        },
        "/PenguinStats/api/v2/zones/{zoneId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zone"
                ],
                "summary": "Get a Zone with ID",
                "deprecated": true,
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Zone ID",
                        "name": "zoneId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/shims.Zone"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "existence": {
                                            "$ref": "#/definitions/models.Existence"
                                        },
                                        "zoneName_i18n": {
                                            "$ref": "#/definitions/models.I18nString"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid or missing zoneId. Notice that this shall be the **string ID** of the zone, instead of the v3 API internally used numerical ID of the zone.",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        },
...
        "/v3/stages": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stage"
                ],
                "summary": "Get All Stages",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "allOf": [
                                    {
                                        "$ref": "#/definitions/models.Stage"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "code": {
                                                "$ref": "#/definitions/models.I18nString"
                                            },
                                            "existence": {
                                                "$ref": "#/definitions/models.Existence"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        },
        "/v3/stages/{stageId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stage"
                ],
                "summary": "Get an Stage with ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stage ID",
                        "name": "stageId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Stage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/models.I18nString"
                                        },
                                        "existence": {
                                            "$ref": "#/definitions/models.Existence"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid or missing stageId. Notice that this shall be the **string ID** of the stage, instead of the internally used numerical ID of the stage.",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        },
        "/v3/zones": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zone"
                ],
                "summary": "Get All Zones",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "allOf": [
                                    {
                                        "$ref": "#/definitions/models.Zone"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "existence": {
                                                "$ref": "#/definitions/models.Existence"
                                            },
                                            "name": {
                                                "$ref": "#/definitions/models.I18nString"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        },
        "/v3/zones/{zoneId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zone"
                ],
                "summary": "Get a Zone with ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Zone ID",
                        "name": "zoneId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Zone"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "existence": {
                                            "$ref": "#/definitions/models.Existence"
                                        },
                                        "name": {
                                            "$ref": "#/definitions/models.I18nString"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid or missing zoneId. Notice that this shall be the **string ID** of the zone, instead of the internally used numerical ID of the zone.",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/errors.PenguinError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errors.Extras": {
            "type": "object",
            "additionalProperties": true
        },
        "errors.PenguinError": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string",
                    "example": "INVALID_REQUEST"
                },
                "extras": {
                    "$ref": "#/definitions/errors.Extras"
                },
                "message": {
                    "type": "string",
                    "example": "invalid request: some or all request parameters are invalid"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "models.Existence": {
            "type": "object",
            "required": [
                "CN",
                "JP",
                "KR",
                "US"
            ],
            "properties": {
                "CN": {
                    "description": "CN: 国服 Mainland China Server (maintained by Hypergryph Network Technology Co., Ltd.)",
                    "$ref": "#/definitions/models.ServerExistence"
                },
                "JP": {
                    "description": "JP: 日服 Japan Server (maintained by Yostar Inc,.)",
                    "$ref": "#/definitions/models.ServerExistence"
                },
                "KR": {
                    "description": "KR: 韩服 Korea Server (maintained by Yostar Limited)",
                    "$ref": "#/definitions/models.ServerExistence"
                },
                "US": {
                    "description": "US: 美服/国际服 Global Server (maintained by Yostar Limited)",
                    "$ref": "#/definitions/models.ServerExistence"
                }
            }
        },
...
        "models.I18nString": {
            "type": "object",
            "required": [
                "en",
                "ja",
                "ko",
                "zh"
            ],
            "properties": {
                "en": {
                    "description": "EN: English (en)",
                    "type": "string"
                },
                "ja": {
                    "description": "JP: 日本語 (ja)",
                    "type": "string"
                },
                "ko": {
                    "description": "KR: 한국어 (ko)",
                    "type": "string"
                },
                "zh": {
                    "description": "ZH: 中文 (zh-CN)",
                    "type": "string"
                }
            }
        },
...
        "models.ServerExistence": {
            "type": "object",
            "required": [
                "exist"
            ],
            "properties": {
                "endTime": {
                    "type": "integer"
                },
                "exist": {
                    "type": "boolean"
                },
                "startTime": {
                    "type": "integer"
                }
            }
        },
        "models.Stage": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code is a map with language code as key and the code of the stage in that language as value.",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "existence": {
                    "description": "Existence is a map with server code as key and the existence of the item in that server as value.",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "extraProcessType": {
                    "description": "ExtraProcessType is the type of extra process that is used in the stage, e.g. \"GACHABOX\".",
                    "type": "string"
                },
                "minClearTime": {
                    "description": "MinClearTime is the minimum time (in milliseconds as a duration) it takes to clear the stage, referencing from prts.wiki",
                    "type": "number"
                },
                "penguinStageId": {
                    "description": "StageID (penguinStageId) is the numerical ID of the stage.",
                    "type": "integer"
                },
                "sanity": {
                    "description": "Sanity is the sanity requirement for a full clear of the stage.",
                    "type": "number"
                },
                "stageId": {
                    "description": "ArkStageID (stageId) is the previously used, string form ID of the stage; in JSON-representation `stageId` is used as key.",
                    "type": "string"
                },
                "stageType": {
                    "description": "StageType is the type of the stage, e.g. \"MAIN\", \"SUB\", \"ACTIVITY\" and \"DAILY\".",
                    "type": "string"
                },
                "zoneId": {
                    "description": "ZoneID is the numerical ID of the zone the stage is in.",
                    "type": "integer"
                }
            }
        },
        "models.Zone": {
            "type": "object",
            "properties": {
                "background": {
                    "description": "Background is the path of the background image of the zone, relative to the CDN endpoint.",
                    "type": "string"
                },
                "category": {
                    "description": "Category of the zone.",
                    "type": "string",
                    "example": "MAINLINE"
                },
                "existence": {
                    "description": "Existence is a map with server code as key and the existence of the item in that server as value.",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "index": {
                    "type": "integer"
                },
                "name": {
                    "description": "Name is a map with language code as key and the name of the item in that language as value.",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "penguinZoneId": {
                    "description": "ZoneID is the numerical ID of the zone.",
                    "type": "integer"
                },
                "type": {
                    "description": "Type of the zone, e.g. \"AWAKENING_HOUR\" or \"VISION_SHATTER\". Optional and only occurres when `category` is \"MAINLINE\".",
                    "type": "string",
                    "example": "AWAKENING_HOUR"
                },
                "zoneId": {
                    "type": "string"
                }
            }
        },
...
        "shims.Zone": {
            "type": "object",
            "properties": {
                "background": {
                    "type": "string"
                },
                "existence": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "stages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "subType": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "zoneId": {
                    "type": "string"
                },
                "zoneIndex": {
                    "type": "integer"
                },
                "zoneName": {
                    "type": "string"
                },
                "zoneName_i18n": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    }
}